Set type to 'object' for union types
Union types couldn't be parsed during swagger generation. The 'fix' for this was somewhat hazy, as it turns out that Swagger/OpenAPI spec doesn't actually support union types in the way that TypeScript offers them.

The fix for now is to set the swagger type of a union type as 'object', which is used to represent arbitrary json values, a.k.a. 'any' or 'dynamic'.