Allow relative locations for OpenApi server URLs
OpenApi 3.0 allows to no specify a scheme/host in the server URL part,
in which case the scheme/host of the location where the document is
served from should then be used by follup consumers.
Until now if nothing special was specified always localhost:3000 was
taken as default.
To not break backwards compatability only enable relative locations when
explicitly specifying "null" as host.

This should most likely be changed with a new major breaking version.

Fixes: https://github.com/lukeautry/tsoa/issues/540