Change default behavior when no host is defined
In OpenAPI 3.0 definitions of an url can be absolute and contain an host, or
relative. This commit changes the bahavior of tsoa to generate a
relative url when no host is defined.

In version 2 of this lib when no host was defined it was as default set
to localhost:3000. The possibility to also generate relative urls
without breaking backwards compatability it was chosen to only generate
such urls when host is set tu 'null'.

For spec versoin 2 the urls where always relative, so this commit
realignes the behavior again.

Fixes #555