IOC: Load ioc module after controllers
If you are using inversify binding decorators, controllers must be
in scope in order to be injected.
In case you're using controllerPathGlobs, you usually are not bringing
them into scope manually, but tsoa does.
If we load the iocModule first, it'll be built before controllers
discovered automatically by the glob are in scope and not able to
provide the controllers.
This applies to authModule aswell I presume