fix: Don't URI encode schema keys when using $ref
This fixes an isssue where using `$` in type names leads to invalid
OpenAPI specs. This would happen because the `refName` property that is
used as the object key for refs and to create `$ref` attributes is URI
encoded on creation, when it only needs to be URI encoded when used in
a `$ref` attribute.

Fixes #1461

Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>